+++ src/CLIException.php
@@ -1,0 +1,21 @@
+<?php
+/**
+ * @namspace      hphio\cli
+ * @name CLIException
+ * Summary: #$END$#
+ *
+ * Date: 2023-08-19
+ * Time: 7:16 AM
+ *
+ * @author        Michael Munger <mj@hph.io>
+ * @copyright (c) 2023 High Powered Help, Inc. All Rights Reserved.
+ */
+
+namespace hphio\cli;
+
+use Exception;
+
+class CLIException extends Exception
+{
+
+}