+++ tests/unit/TreeTestingTraits.php
@@ -1,16 +1,19 @@
 <?php
-namespace hphio\cli;
+namespace Tests\unit;
+
+use hphio\cli\BaseCli;
 
 /**
  * {CLASS SUMMARY}
  *
  * Date: 9/17/18
  * Time: 10:13 PM
+ *
  * @author Michael Munger <mj@hph.io>
  */
 trait TreeTestingTraits
 {
-    public function getSampleTreeStructure()
+    public function getSampleTreeStructure(): array
     {
         $commands = ["user password reset", "user password invalidate", "user deactivate", "user activate", "show help", "show users", "foo bar", 'baz bar foo', 'baz barf zoo'];
         $treeStructure = [];