+++ CHANGELOG.md
@@ -1,4 +1,24 @@
-# Unrelease
+# Unreleased
+
+# v2.0.0
+## Added
+- Added setCommandText and setHelpText methods to AbstractCommand.
+
+## Changed
+- Updated CommandInterface to require setCommand and setHelp methods to set the command and help text.
+- Changed / fixed namespaces.
+
+## Deprecated
+- Nothing yet.
+
+## Removed
+- Nothing yet.
+## Fixed
+- Nothing yet.
+## Security
+- Nothing yet.
+
+v1.x
 
 ## Added
 - This changelog.